webui: split poms into pages and operations subdirectories
Amp-Thread-ID: https://ampcode.com/threads/T-019c7cdd-5005-773a-b698-ac98cc5dc9bb
