+++ webui/test/e2e/common/setupInfra.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect } from "@playwright/test";
-import { SetupPage } from "../poms/setupPage";
-import { LoginPage } from "../poms/loginPage";
-import { RepositoriesPage } from "../poms/repositoriesPage";
+import { SetupPage } from "../poms/pages/setupPage";
+import { LoginPage } from "../poms/pages/loginPage";
+import { RepositoriesPage } from "../poms/pages/repositoriesPage";
 import { STORAGE_STATE_PATH, writeCredentials, ensureStorageDir } from "../credentialsFile";
 
 test("setup lakeFS and save auth state", async ({ page }) => {
