+++ webui/test/e2e/poms/operations/revertOperations.ts
@@ -1,5 +1,5 @@
 import { Page } from "@playwright/test";
-import { TIMEOUT_NAVIGATION } from "../timeouts";
+import { TIMEOUT_NAVIGATION } from "../../timeouts";
 
 export class RevertOperations {
     constructor(private page: Page) {}
